+++ locals-agic.tf
@@ -16,8 +16,15 @@ locals {
     zones                               = ["1", "2", "3"]
     policy_name                         = "AppGwSslPolicy20170401S"
     enabled_waf                         = false
-    app_gateway_tags                    = local.default_tags
-    ip_tags                             = local.default_tags
+    file_upload_limit_mb                = "100"
+    max_request_body_size_kb            = "128"
+    request_body_check                  = true
+    rule_set_type                       = "OWASP"
+    rule_set_version                    = "3.0"
+    firewall_mode                       = "Detection"
+
+    app_gateway_tags = local.default_tags
+    ip_tags          = local.default_tags
     appgw_backend_http_settings = [{
       name                  = "dummy"
       cookie_based_affinity = "Disabled"
@@ -63,6 +70,7 @@ locals {
       port = 80
     }]
     ssl_certificates_configs = []
+    identity                 = var.enable_appgw_msi ? azurerm_user_assigned_identity.appgw_assigned_identity[0].id : null
   }
 
   appgw_settings                 = merge(local.appgw_default_settings, var.appgw_settings)